secured endpoints

+++ part3/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -17,16 +18,37 @@ review_model = api.model('Review', {
 
 @api.route('/')
 class ReviewList(Resource):
+    @jwt_required()
+    @api.doc(security='Bearer')
     @api.expect(review_model, validate=True)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(409, 'User has already reviewed this place')
     def post(self):
+        token_user_id = get_jwt_identity()
         """Register a new review"""
         review_data = api.payload
+
+        # we need to get the place 1st to check the owner id
+        place_data = facade.get_place(review_data['place_id'])
+        if not place_data:
+            return {'message': 'Place not found'}, 404
+
+        # check user (from jwt) is not the owner of the place
+        if token_user_id == place_data["owner"]:
+            return {'message': 'Unauthorized'}, 401
+
+        # Check if the user has already reviewed this place
+        existing_review = facade.get_user_review_for_place(
+            token_user_id,
+            review_data['place_id'])
+        # 409 is for conflict
+        if existing_review:
+            return {'message': 'User has already reviewed this place'}, 409
+
         new_review = facade.create_review(review_data)
         if not new_review:
             return {'error': 'Invalid input data'}, 400
-
         return new_review, 201
 
     @api.response(200, 'List of reviews retrieved successfully')
@@ -47,22 +69,44 @@ class ReviewResource(Resource):
             return {'error': 'Review not found'}, 404
         return review, 200
 
+    @jwt_required()
+    @api.doc(security='Bearer')
     @api.expect(review_model)
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
     def put(self, review_id):
+        token_user_id = get_jwt_identity()
         """Update a review's information"""
         review_data = api.payload
+
+        existing_review = facade.get_review(review_id)
+        if not existing_review:
+            return {'error': 'Review not found or invalid data'}, 404
+
+        # check user (from jwt) is the original author of the review
+        if not token_user_id == existing_review["user_id"]:
+            return {'message': 'Unauthorized'}, 401
         updated_review = facade.update_review(review_id, review_data)
         if not updated_review:
             return {'error': 'Review not found or invalid data'}, 404
 
         return updated_review, 200
 
+    @jwt_required()
+    @api.doc(security='Bearer')
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
     def delete(self, review_id):
+        token_user_id = get_jwt_identity()
+
+        existing_review = facade.get_review(review_id)
+        if not existing_review:
+            return {'error': 'Review not found or invalid data'}, 404
+
+        # check user (from jwt) is the original author of the review
+        if not token_user_id == existing_review["user_id"]:
+            return {'message': 'Unauthorized'}, 401
         """Delete a review"""
         success = facade.delete_review(review_id)
         if not success:
